dockerize python without db
