+++ app/Out.json
@@ -1,0 +1,1 @@
+{"Tractor_ID":{"0":"2P-802","1":"2P-803","2":"2P-804","3":"2P-805","4":"2P-806","5":"2P-807","6":"2P-808","7":"2P-809","8":"2P-103","9":"2P-104","10":"2P-105","11":"2P-106","12":"2P-710","13":"2P-717","14":"2P-724","15":"2P-726","16":"2P-728","17":"2P-730","18":"2P-731","19":"2P-733","20":"2P-736","21":"2P-737"},"Tractor_ava":{"0":"No","1":"Yes","2":"Yes","3":"Yes","4":"Yes","5":"Yes","6":"No","7":"Yes","8":"Yes","9":"Yes","10":"No","11":"No","12":"Yes","13":"Yes","14":"No","15":"No","16":"Yes","17":"Yes","18":"Yes","19":"No","20":"No","21":"No"},"Tractor_inuse":{"0":"No","1":"No","2":"No","3":"No","4":"No","5":"Yes","6":"No","7":"Yes","8":"No","9":"No","10":"No","11":"No","12":"No","13":"No","14":"No","15":"No","16":"No","17":"No","18":"Yes","19":"No","20":"No","21":"No"},"Tractor_size":{"0":"Large","1":"Large","2":"Medium","3":"Medium","4":"Large","5":"Large","6":"Large","7":"Large","8":"Medium","9":"Medium","10":"Medium","11":"Medium","12":"Large","13":"Medium","14":"Large","15":"Large","16":"Large","17":"Large","18":"Large","19":"Large","20":"Large","21":"Medium"},"Tractor_type":{"0":"ATT Towbarless Pushback","1":"ATT Towbarless Pushback","2":"ATT Towbarless Pushback","3":"ATT Towbarless Pushback","4":"ATT Towbarless Pushback","5":"ATT Towbarless Pushback","6":"ATT Towbarless Pushback","7":"ATT Towbarless Pushback","8":"ACT Conventional Pushback","9":"ACT Conventional Pushback","10":"ACT Conventional Pushback","11":"ACT Conventional Pushback","12":"ACT Conventional Pushback","13":"ACT Conventional Pushback","14":"ACT Conventional Pushback","15":"ACT Conventional Pushback","16":"ACT Conventional Pushback","17":"ACT Conventional Pushback","18":"ACT Conventional Pushback","19":"ACT Conventional Pushback","20":"ACT Conventional Pushback","21":"ACT Conventional Pushback"},"Tractor_brand":{"0":"Goldhofer","1":"Goldhofer","2":"Kalmar","3":"Kalmar","4":"Goldhofer","5":"Goldhofer","6":"Goldhofer","7":"Goldhofer","8":"Schopf","9":"Schopf","10":"Schopf","11":"Schopf","12":"Schopf","13":"Schopf","14":"Fresia","15":"Fresia","16":"Fresia","17":"Schopf","18":"Schopf","19":"Schopf","20":"Schopf","21":"Schopf"},"Tractor_Prefer":{"0":null,"1":null,"2":null,"3":null,"4":"Narrow","5":"Narrow","6":"Narrow","7":"Narrow","8":null,"9":null,"10":null,"11":null,"12":null,"13":null,"14":null,"15":null,"16":null,"17":null,"18":null,"19":null,"20":null,"21":null},"Last_location":{"0":"G7","1":"E8","2":"E6","3":"104","4":"118","5":"507","6":"G7","7":"C4","8":"110","9":"B1","10":"G7","11":"G7","12":"E2","13":"C8","14":"G7","15":"G7","16":"F5","17":"120","18":"C8","19":"G7","20":"G7","21":"G7"},"Flight_counter":{"0":0,"1":18,"2":35,"3":4,"4":43,"5":43,"6":0,"7":44,"8":21,"9":36,"10":0,"11":0,"12":43,"13":37,"14":0,"15":0,"16":43,"17":43,"18":44,"19":0,"20":0,"21":0},"Prev_flight":{"0":null,"1":"JL 034","2":"LJ 012","3":"WE 570","4":"TG 345","5":"W5 050","6":null,"7":"TG 329","8":"WE 267","9":"WE 211","10":null,"11":null,"12":"ZH 9002","13":"TG 341","14":null,"15":null,"16":"VN 606","17":"TG 325","18":"TG 337","19":null,"20":null,"21":null},"Des_flight":{"0":null,"1":"TG 682","2":"KE 652","3":"WE 291","4":"TG 682","5":"TG 658","6":null,"7":"TG 331","8":"WE 211","9":"WE 341","10":null,"11":null,"12":"NH 850","13":"TG 337","14":null,"15":null,"16":"5X 145","17":"TG 351","18":"TG 656","19":null,"20":null,"21":null},"Des_Bay":{"0":null,"1":"126","2":"F6","3":"A4","4":"126","5":"C2","6":null,"7":"C6","8":"B1","9":"E6","10":null,"11":null,"12":"E7","13":"C8","14":null,"15":null,"16":"522","17":"114","18":"C10","19":null,"20":null,"21":null},"Distance":{"0":null,"1":2910.047,"2":564.035,"3":725.37,"4":2941.736,"5":1263.43,"6":null,"7":1416.43,"8":1591.135,"9":564.035,"10":null,"11":null,"12":747.84,"13":76.5,"14":null,"15":null,"16":2058.387,"17":2967.649,"18":1569.43,"19":null,"20":null,"21":null},"E90":{"0":null,"1":null,"2":"Y","3":"Y","4":null,"5":null,"6":null,"7":null,"8":null,"9":null,"10":null,"11":null,"12":null,"13":null,"14":null,"15":null,"16":null,"17":null,"18":null,"19":null,"20":null,"21":null},"A319":{"0":null,"1":null,"2":"Y","3":"Y","4":"Y","5":"Y","6":"Y","7":"Y","8":"Y","9":"Y","10":"Y","11":"Y","12":"Y","13":"Y","14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":"Y"},"A320":{"0":null,"1":null,"2":"Y","3":"Y","4":"Y","5":"Y","6":"Y","7":"Y","8":"Y","9":"Y","10":"Y","11":"Y","12":"Y","13":"Y","14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":"Y"},"A321":{"0":null,"1":null,"2":"Y","3":"Y","4":"Y","5":"Y","6":"Y","7":"Y","8":"Y","9":"Y","10":"Y","11":"Y","12":"Y","13":"Y","14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":"Y"},"A330":{"0":"Y","1":"Y","2":null,"3":null,"4":"Y","5":"Y","6":"Y","7":"Y","8":null,"9":null,"10":null,"11":null,"12":"Y","13":null,"14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":null},"A340":{"0":"Y","1":"Y","2":null,"3":null,"4":"Y","5":"Y","6":"Y","7":"Y","8":null,"9":null,"10":null,"11":null,"12":"Y","13":null,"14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":null},"A350":{"0":"Y","1":"Y","2":null,"3":null,"4":"Y","5":"Y","6":"Y","7":"Y","8":null,"9":null,"10":null,"11":null,"12":"Y","13":null,"14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":null},"A359":{"0":"Y","1":"Y","2":null,"3":null,"4":"Y","5":"Y","6":"Y","7":"Y","8":null,"9":null,"10":null,"11":null,"12":"Y","13":null,"14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":null},"A380":{"0":"Y","1":"Y","2":null,"3":null,"4":null,"5":null,"6":null,"7":null,"8":null,"9":null,"10":null,"11":null,"12":"Y","13":null,"14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":null},"B737":{"0":null,"1":null,"2":"Y","3":"Y","4":"Y","5":"Y","6":"Y","7":"Y","8":"Y","9":"Y","10":"Y","11":"Y","12":"Y","13":"Y","14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":"Y"},"B738":{"0":null,"1":null,"2":"Y","3":"Y","4":"Y","5":"Y","6":"Y","7":"Y","8":"Y","9":"Y","10":"Y","11":"Y","12":"Y","13":"Y","14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":"Y"},"B747":{"0":"Y","1":"Y","2":null,"3":null,"4":null,"5":null,"6":null,"7":null,"8":null,"9":null,"10":null,"11":null,"12":"Y","13":null,"14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":null},"B748":{"0":"Y","1":"Y","2":null,"3":null,"4":null,"5":null,"6":null,"7":null,"8":null,"9":null,"10":null,"11":null,"12":null,"13":null,"14":null,"15":null,"16":null,"17":null,"18":null,"19":null,"20":null,"21":null},"B757":{"0":null,"1":null,"2":null,"3":null,"4":"Y","5":"Y","6":"Y","7":"Y","8":null,"9":null,"10":null,"11":null,"12":"Y","13":null,"14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":null},"B767":{"0":null,"1":null,"2":null,"3":null,"4":"Y","5":"Y","6":"Y","7":"Y","8":null,"9":null,"10":null,"11":null,"12":"Y","13":null,"14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":null},"B777":{"0":"Y","1":"Y","2":"Y","3":"Y","4":"Y","5":"Y","6":"Y","7":"Y","8":null,"9":null,"10":null,"11":null,"12":"Y","13":"Y","14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":null},"B787":{"0":"Y","1":"Y","2":null,"3":null,"4":"Y","5":"Y","6":"Y","7":"Y","8":null,"9":null,"10":null,"11":null,"12":"Y","13":null,"14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":null},"B788":{"0":"Y","1":"Y","2":null,"3":null,"4":"Y","5":"Y","6":"Y","7":"Y","8":null,"9":null,"10":null,"11":null,"12":"Y","13":null,"14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":null},"B789":{"0":"Y","1":"Y","2":null,"3":null,"4":"Y","5":"Y","6":"Y","7":"Y","8":null,"9":null,"10":null,"11":null,"12":"Y","13":null,"14":"Y","15":"Y","16":"Y","17":"Y","18":"Y","19":"Y","20":"Y","21":null}}
