+++ app/fltpair.txt
@@ -1,0 +1,1 @@
+flight;Tractor
