+++ requirements.txt
@@ -1,0 +1,6 @@
+numpy
+pandas==2.1.0
+psycopg2
+requests
+scipy
+openpyxl
